Make checker dialog colors configurable
1. Current line in checker list of messages
2. Highlighted word in message, e.g. "he" in Jeebies message
3. Error prefix, e.g. type of error in Curly Quote Check

Doesn't let you configure bg or fg of a color if it would have
no effect, e.g. error prefix is fg only, and main window
selected color when window not focused (inactive) is bg
only.

Also miscellaneous code re-org, some of which was needed,
and some just to improve it.
